+++ gamemodes/terrortown/gamemode/client/cl_main.lua
@@ -519,7 +519,7 @@ function GM:ClearClientState()
 
     client:SetTargetPlayer(nil)
 
-    VOICE.InitBattery()
+    voicebattery.InitBattery()
 
     local plys = playerGetAll()