+++ CHANGELOG.md
@@ -39,6 +39,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
   - Added a submenu to issue basic commands
 - Added a new `gameloop` module that contains all functions related to the round structure (by @Tim Goll)
 - Added a loadingscreen that hides the visible and audible lag introduced by the map cleanup on round change (by @TimGoll)
+- Added a voicebattery module that handles the voice battery (by @TimGoll)
 
 ### Changed